Move bin logic to lib
